+++ rocketgram/api/chat_member.py
@@ -3,6 +3,7 @@
 # Rocketgram is released under the MIT License (see LICENSE).
 
 
+import warnings
 from dataclasses import dataclass
 from datetime import datetime
 from typing import Optional
@@ -29,7 +30,7 @@ class ChatMember:
     can_post_messages: Optional[bool]
     can_edit_messages: Optional[bool]
     can_delete_messages: Optional[bool]
-    can_manage_voice_chats: Optional[bool]
+    can_manage_video_chats: Optional[bool]
     can_invite_users: Optional[bool]
     can_restrict_members: Optional[bool]
     can_pin_messages: Optional[bool]
@@ -40,6 +41,12 @@ class ChatMember:
     can_send_polls: Optional[bool]
     can_send_other_messages: Optional[bool]
     can_add_web_page_previews: Optional[bool]
+
+    @property
+    def can_manage_voice_chats(self) -> Optional[bool]:
+        warnings.warn("This field is deprecated. Use `can_manage_video_chats` instead.", DeprecationWarning)
+
+        return self.can_manage_video_chats
 
     @classmethod
     def parse(cls, data: dict) -> Optional['ChatMember']: