+++ rocketgram/api/message_type.py
@@ -40,10 +40,15 @@ class MessageType(EnumAutoName):
     connected_website = auto()
     passport_data = auto()
     proximity_alert_triggered = auto()
-    voice_chat_scheduled = auto()
-    voice_chat_started = auto()
-    voice_chat_ended = auto()
-    voice_chat_participants_invited = auto()
+    video_chat_scheduled = auto()
+    video_chat_started = auto()
+    video_chat_ended = auto()
+    video_chat_participants_invited = auto()
     message_auto_delete_timer_changed = auto()
     web_app_data = auto()
     unknown = auto()
+
+    voice_chat_scheduled = video_chat_scheduled  # deprecated
+    voice_chat_started = video_chat_started  # deprecated
+    voice_chat_ended = video_chat_ended  # deprecated
+    voice_chat_participants_invited = video_chat_participants_invited  # deprecated