+++ rocketgram/api/video_chat_ended.py
@@ -3,22 +3,29 @@
 # Rocketgram is released under the MIT License (see LICENSE).
 
 
+import warnings
 from dataclasses import dataclass
 from typing import Dict, Optional
 
 
 @dataclass(frozen=True)
-class VoiceChatEnded:
+class VideoChatEnded:
     """\
-    Represents VoiceChatEnded object:
-    https://core.telegram.org/bots/api#voicechatended
+    Represents VideoChatEnded object:
+    https://core.telegram.org/bots/api#videochatended
     """
 
     duration: int
 
     @classmethod
-    def parse(cls, data: Optional[Dict]) -> Optional['VoiceChatEnded']:
+    def parse(cls, data: Optional[Dict]) -> Optional['VideoChatEnded']:
         if data is None:
             return None
 
         return cls(data['duration'])
+
+
+def VoiceChatEnded(*args, **kwargs):
+    warnings.warn("This class is deprecated. Use `VideoChatEnded` instead.", DeprecationWarning)
+
+    return VideoChatEnded(*args, **kwargs)